Compatiblity: ensure compatibility with Jetpack 11.8
In Jetpack 11.8 or higher, the plugin includes Mastodon as a valid Social Menu option. Let's fallback to Jetpack's implementation in this case.

This only fallbacks to Jetpack for the Social Menu logic; it doesnt touch the block theme feature, since that's not supported by Jetpack.

For reference, here is Jetpack's PR:
https://github.com/Automattic/jetpack/pull/28175